Added features for adding dates, changing them and get all the available commands

+++ bot/internal/models/bridge/chat.go
@@ -1,9 +1,9 @@
 package bridge
 
 type Chat struct {
-	chatId int
+	chatId int64
 }
 
-func NewChat(chatId int) *Chat {
+func NewChat(chatId int64) *Chat {
 	return &Chat{chatId: chatId}
 }
