+++ bot/internal/models/bridge/user.go
@@ -1,11 +1,11 @@
 package bridge
 
 type User struct {
-	userId      int
+	userId      int64
 	date        string
 	currentChat Chat
 }
 
-func NewUser(id int, date string, chat Chat) *User {
+func NewUser(id int64, date string, chat Chat) *User {
 	return &User{userId: id, date: date, currentChat: chat}
 }
